Revert "Revert "7:28 곽건:skill모듈을 만드는 중""
This reverts commit faa3aaa62c76c7b4f67a7a2fef0dd8b27ca37d84.
뭔가 잘못 건드렸다
